+++ BackAppCore/Sources/BackAppCore/BakingRecipeItems/StepItem.swift
@@ -12,8 +12,8 @@ import BakingRecipeFoundation
 public class StepItem: Item {
     public var step: Step
     
-    public init(id: Int, step: Step) {
+    public init(id: Int64, step: Step) {
         self.step = step
-        super.init(id: id)
+        super.init(id: Int(id))
     }
 }
